requirements.txt
azure-search-documents==11.6.0b9
azure-storage-blob==12.24.0
azure-ai-documentintelligence>=1.0.0
openai>=1.0.0
openpyxl==3.1.5
python-dotenv==1.0.1

# API layer
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# Foundry provider (only needed when CHAT_PROVIDER=foundry)
# Install with --no-deps on networks that block zipp/importlib-metadata:
#   pip install agent-framework-core agent-framework-openai agent-framework-foundry \
#               azure-ai-inference azure-ai-projects azure-identity \
#               opentelemetry-api importlib-metadata msal --no-deps
agent-framework-core>=1.1.0
agent-framework-openai>=1.1.0
agent-framework-foundry>=1.1.0
azure-ai-inference>=1.0.0b9
azure-ai-projects>=2.1.0
azure-identity>=1.15.0
opentelemetry-api>=1.39.0
importlib-metadata>=6.0
msal>=1.30.0
